fix: made each section in the about into it's own components and seperate files
