+++ frontend/src/components/about-sections/AboutHero.jsx
@@ -1,0 +1,12 @@
+export default function AboutHero() {
+  return (
+    <section className="text-center mb-20">
+      <h1 className="text-5xl md:text-6xl font-bold mb-6">
+        About <span className="text-primary">WanderBite</span>
+      </h1>
+      <p className="text-xl text-muted-foreground max-w-3xl mx-auto">
+        Connecting food lovers with authentic local experiences around the world
+      </p>
+    </section>
+  )
+}
